--- src/ChildrenComponents.tsx
@@ -1,5 +1,0 @@
-import { ProviderProps } from './types/types';
-
-export const ChildrenComponents = ({ children }: ProviderProps) => {
-  return <>{children}</>;
-};
